refactor:Separate declarations #1
